+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Make `Pledge` polyfill for `Promise`
 * Add `Pledge.after(n, val)`, which returns a pledge that resolves after n milliseconds
 * Add `Pledge#race(other_pledge)`
+* Add new `Array#findByPath(obj)` signature
 
 ## 0.5.11 (2018-07-14)